v4: add option 121
This adds support for Classless Static Routes in both the encoder and
the decoder.

For ease of use, a new dependency on the ipnet crate is added to provide
an IPv4 Network structure and avoid carrying `Ipv4` + `u8`, and having the
prefix length checked.

When the `serde` feature of the crate is enabled, the corresponding
feature of `ipnet` is also enabled to have full serde compatibility.